+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 dev:
   - reduce time spent verifying account names
   - ensure that attestations complete on Vouch's first ever epoch
+  - allow Vouch to start with some consensus nodes unavailable
 
 1.8.0:
   - support Deneb